@@ -1,14 +1,14 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 32
+    buildToolsVersion '30.0.3'
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 27
-        versionCode 2
-        versionName "1.1"
+        targetSdkVersion 32
+        versionCode 3
+        versionName "1.2"
 
     }
     buildTypes {
@@ -21,5 +21,6 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.annotation:annotation:1.3.0'
+    implementation 'androidx.interpolator:interpolator:1.0.0'
 }
