@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 32
+    buildToolsVersion '30.0.3'
     defaultConfig {
         applicationId "com.github.cdflynn.android.checksample"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
     }
@@ -20,6 +20,6 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
     implementation project(':checkview')
+    implementation 'androidx.appcompat:appcompat:1.4.1'
 }
