@@ -1,9 +1,9 @@
 package com.github.cdflynn.android.checksample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
+import androidx.appcompat.app.AppCompatActivity;
 import nl.schmit.animationView.CrossView;
 import nl.schmit.animationView.ExclamationView;
 
